+++ R/as_prediction_clust.R
@@ -43,10 +43,10 @@ as_prediction_clust.PredictionClust = function(x, ...) {
 #' @export
 as_prediction_clust.data.frame = function(x, ...) {
   assert_names(names(x), must.include = c("row_ids", "partition"))
-  prob_cols = setdiff(names(x), c("row_ids", "partition"))
+  prob_cols = setdiff(names(x), c("row_ids", "partition", "weights"))
   if (length(prob_cols) > 0L && !all(startsWith(prob_cols, "prob."))) {
     error_input(
-      "Table may only contain columns 'row_ids', 'partition' as well as columns prefixed with 'prob.' for class probabilities." # nolint
+      "Table may only contain columns 'row_ids', 'partition', 'weights' as well as columns prefixed with 'prob.' for class probabilities." # nolint
     )
   }